test: add unit test placeholder for "main", no tests implemented yet
to compile main into a unit test required addition of a lot
of MCU registers as variables in the test framework.

this change adds the main module to the code coverage so it
will be accounted in the code coverage report. It is a future
task to implement the unit test for the main module.